% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hello}
\alias{hello}
\title{Say Hello from Scala}
\usage{
hello(sc)
}
\arguments{
\item{sc}{spark connection}
}
\value{
character hello message
}
\description{
Say Hello from Scala
}
\examples{
\dontrun{
library(sparklyr)
library(sparkhello)

sc <- spark_connect(master = "local")

sparkhello_hello(sc)

spark_disconnect(sc)
}
}
